Remove redundant join/split
I don't know how often this code runs.  Seems unnecessary to
`.join` on line 142 only to split and rejoin again in `indent`.

`indent` now checks for the default case and should behave as before.
If the `split` is not default it will treat unindented as an array.
It will then rejoin with the `split` parameter.